Added better logging in sliced gcode files
I was struggling to remember which material profile I used for which particular material. So if you open the .gcode file with a text editor (I recommend Notepad++) it will now have at the top of the .gcode file which print, filament and printer profiles were used to slice that particular file.